refactor envelope factories to avoid circular dependencies
