+++ wsdiscovery/actions/__init__.py
@@ -2,11 +2,11 @@
 The actions subpackage provides WS-Discovery action message construction and parsing.
 """
 
-from .bye import NS_ACTION_BYE, createByeMessage, parseByeMessage
-from .hello import NS_ACTION_HELLO, createHelloMessage, parseHelloMessage
-from .probe import NS_ACTION_PROBE, createProbeMessage, parseProbeMessage
-from .probematch import NS_ACTION_PROBE_MATCH, createProbeMatchMessage, parseProbeMatchMessage
+from .bye import NS_ACTION_BYE, constructBye, createByeMessage, parseByeMessage
+from .hello import NS_ACTION_HELLO, constructHello, createHelloMessage, parseHelloMessage
+from .probe import NS_ACTION_PROBE, constructProbe, createProbeMessage, parseProbeMessage
+from .probematch import NS_ACTION_PROBE_MATCH, constructProbeMatch, createProbeMatchMessage, parseProbeMatchMessage
 from .probematch import ProbeResolveMatch
-from .resolve import NS_ACTION_RESOLVE, createResolveMessage, parseResolveMessage
-from .resolvematch import NS_ACTION_RESOLVE_MATCH, createResolveMatchMessage, parseResolveMatchMessage
+from .resolve import NS_ACTION_RESOLVE, constructResolve, createResolveMessage, parseResolveMessage
+from .resolvematch import NS_ACTION_RESOLVE_MATCH, constructResolveMatch, createResolveMatchMessage, parseResolveMatchMessage
